Transfer Owned Listing logic to a protected route, rename protected routes, fix chat card heading issue.
