+++ growers_market.client/src/Components/ChatCard/ChatCard.css
@@ -8,7 +8,7 @@
     padding: 10px;
 }
 
-    .chat-card h3 {
+    .chat-card h4 {
         font-size: 2rem;
         color: var(--color-primary-dark);
         margin-bottom: 10px;
